Use priority instead of weight to determine execution order
The previous commit made all tasks of the same weight execute in order
of niceness, but lighter items were still executed first. That obviously
defeats the purpose of having a priority parameter.

Now, the least nice items are scheduled first. At the same priority
level, execution order is now FIFO. Weight is no longer a factor in
execution order.

The execution queue is now implemented using a single JavaScript array,
rather than an array of arrays by weight. Performance will be noticeably
worse when scheduling a very (very) large number of tasks, especially
when they have diverse weights, but probably better with a large number
of semaphores with few tasks due to fewer array allocations. If there are
users for whom this is a concern, a good heap-based priority queue could
be added as a dependency.